backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
dj-database-url>=2.2
psycopg2-binary>=2.9
django-cors-headers>=4.4
gunicorn>=22
django-environ>=0.11
drf-spectacular
